#FEM-1437
* Added public modifiers for AdEvents to be able to integrate other ad players and posting ad events.
* Moved `PKPlaybackInfo` + `PKAdInfo` + `PKAdCuePoints` to "/Events/EventsPayload"
