+++ Classes/Events/EventsPayloads/PKAdCuePoints.swift
@@ -12,7 +12,7 @@ import Foundation
     
     @objc public private(set) var cuePoints: [TimeInterval]
     
-    init(cuePoints: [TimeInterval]) {
+    public init(cuePoints: [TimeInterval]) {
         self.cuePoints = cuePoints.sorted() // makes sure array is sorted
     }
